twitter_analysis/requirements.txt
TwitterAPI
google-cloud-language
pandas
matplotlib